move validate persistence to utils
