+++ ahnlich/db/src/main.rs
@@ -1,7 +1,7 @@
 use clap::Parser;
 
-use std::{error::Error, os::unix::fs::MetadataExt, path::PathBuf};
-use utils::server::AhnlichServerUtils;
+use std::error::Error;
+use utils::{cli::validate_persistence, server::AhnlichServerUtils};
 
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn Error>> {
@@ -18,25 +18,5 @@ async fn main() -> Result<(), Box<dyn Error>> {
             server.start().await?;
         }
     }
-    Ok(())
-}
-
-fn validate_persistence(
-    allocated_size: usize,
-    persistence_file: Option<&PathBuf>,
-) -> Result<(), String> {
-    if let Some(path_file) = persistence_file {
-        let path = path_file.as_path();
-        if path.is_file() {
-            let metadata = std::fs::metadata(path).map_err(|err| err.to_string())?;
-            if (allocated_size / metadata.size() as usize) < 2 {
-                return Err(
-                    "Allocated memory should be more than two times your persistence_file size"
-                        .to_string(),
-                );
-            }
-        }
-    }
-
     Ok(())
 }
